Allow the decorators to alias properties
**Overview**
@Getters, @Setters and @Properties previously took a property name and
assumed that this property would be named the same on the underlying
model as it was on the GraphQL type object. In real world applications,
this is not always the case. Previously code looked like this

```javascript
@Getters('name')
class MyClass extends GQLBase { }
```

This would allow instances of MyClass to reference .name which would
then directly map to .model.name.

The new code now allows you to alias this by doing the following

```javascript
@Getters(['name', 'firstName'])
class MyClass extends GQLBase { }
```

Now a reference to .name would actually map to .model.firstName

**Changes**
 * Allow the passing of an array for each property where the first value is the name of the property on the GQLBase class and the second is the name of the underlying model property.
 * Updated unit tests