Auto-props, ModuleParser, Unit Tests, Oh My!
Lattice embraces the usage of decorators. The one that is most often
used next to @Schema is @Properties. @Properties allows you to easily
specify 1:1 mapping between the type field value and the model property
of the same name it is backed by. While allowing other more advanced
usages, there is no real need for developers to specify @Properties for
fields that are simply going to be 1:1 mappings when that can be done
automatically for you.

Enter auto-props. Auto-props runs on an instance when the instance is
instantiated. If there is no defined getter for the data in a type as
per its schema, one that points to a property in the model object of the
same name is autogenerated for you.

`ModuleParser` has provided some headaches when it comes to debugging
types that are misbehaving. In an effort to make its usage a bit more
expected, a few non-breaking changes have been made in this release.

 1. First of all, module parser now only picks up files with
 registered extensions. These, by default, are .js, .jsx, .ts and .tsx.
 More, however can be specified in your package.json options for
 lattice. See below for details

 2. Errors in parsing are now captured and can be printed to the console
 or introspected after a ModuleParser run. The mapping between absolute
 file path and the thrown Error object are stored in the `.skipped`
 instance property of each ModuleParser instance. These can be printed
 using `.printSkipped()`

 3. There is a new preference for ModuleParser, optionally set in
 `package.json`, that will cause any errors that occurred to throw a
 parsing error and hopefully stop everything after each of the
 encountered errors are printed to output.

You can now provide some optional preference overrides in
`package.json`. The default values are as follows

```javascript
{
  ModuleParser: {
    extensions: ['.js', '.jsx', '.ts', '.tsx'],
    failOnError: false
  }
}
```

They can be specified or overridden under the "lattice" property

```json
{
  ...
  "lattice": {
    "ModuleParser": {
      "extensions": ['.zs'],
      "failOnError": true
    }
  }
}
```